Working on some guts. Also renamed.
